Print (most of) .eh_frame CFI information in src/examples/dwarfdump.rs
This does most of the information that the canonical dwarfdump does. Still
remaining are things like symbolicating the start addresses in address ranges,
turning register numbers into the target architecture's register names, and
printing augmentation data. None of this is impossible to implement for any
reason right now, I just haven't done it.

Fixes half of #116